requirements.txt
python-docx
pypdf
python-pptx